@@ -3,18 +3,18 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.5.0",
-        "@testing-library/user-event": "^7.2.1",
-        "classnames": "^2.2.6",
-        "primeflex": "^2.0.0",
-        "primeicons": "^4.0.0",
-        "primereact": "^5.0.0-rc.2",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "react-scripts": "3.4.3",
-        "react-transition-group": "^4.4.1",
-        "react-world-flags": "^1.4.0"
+        "@testing-library/jest-dom": "^6.4.2",
+        "@testing-library/react": "^14.2.1",
+        "@testing-library/user-event": "^14.5.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-scripts": "^5.0.1",
+        "react-transition-group": "^4.4.5",
+        "react-world-flags": "^1.6.0",
+        "classnames": "^2.5.1",
+        "primeflex": "^3.3.1",
+        "primeicons": "^6.0.1",
+        "primereact": "^10.5.1"
     },
     "scripts": {
         "start": "react-scripts start",
