@@ -1,7 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
-import PersonManager from './PersonManager';
+import PersonManager from './components/PersonManager';
 // import TestManager from './TestManager';
 
 import './index.css';
@@ -17,5 +16,6 @@ import './styles/widths.css';
 
 import 'primeflex/primeflex.css';
 
-ReactDOM.render(<PersonManager />, document.getElementById('root'));
-// ReactDOM.render(<TestManager />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<PersonManager />);
